added GSI to routeLoader
