+++ containers/backend/src/api/routes/auth/authRoutesIndex.js
@@ -2,12 +2,14 @@ const {
 	signupSchema,
 	signinSchema,
 	logoutSchema,
+	googleSchema
  } = require('../../schemas/auth');
  
 const { 
 	signupHandler,
 	signinHandler,
 	logoutHandler,
+	googleHandler
  } = require('../../handlers/auth');
 
 module.exports = async function (fastify, options) {
@@ -15,6 +17,5 @@ module.exports = async function (fastify, options) {
   fastify.post('/api/auth/signup', { schema: signupSchema }, signupHandler);
   fastify.post('/api/auth/signin', { schema: signinSchema }, signinHandler);
   fastify.post('/api/auth/logout', { schema: logoutSchema }, logoutHandler);
-
-  fastify.register(require('./googleSignUp'));
+  fastify.post('/api/auth/google', { schema: googleSchema }, googleHandler);
 };
