+++ containers/backend/src/api/schemas/auth.js
@@ -177,8 +177,67 @@ const logoutSchema = {
 	}
   };
 
+const googleSchema = {
+	description: 'Create a new user account from the data provided by the user via the frontend. The password is received in plain text and is hashed before being stored in the database.',
+	tags: ['authentication'],
+	body: {
+	  type: 'object',
+	  required: ['username', 'email', 'password'],
+	  properties: {
+		username: { type: 'string', description: 'Unique username' },
+		email: { type: 'string', format: 'email', description: 'User email address' },
+		password: { type: 'string', description: 'User password' }
+	  }
+	},
+	response: {
+	  201: {
+		description: 'Successful registration',
+		type: 'object',
+		properties: {
+		  success: { type: 'boolean' },
+		  message: { type: 'string' },
+		  userId: { type: 'number', description: 'ID of the newly registered user' },    // <--- ADD THIS
+		  username: { type: 'string', description: 'Username of the new user' }, // <--- ADD THIS
+		  email: { type: 'string', description: 'Email of the new user' }
+		},
+		example: {
+		  success: true,
+		  message: 'User registered successfully',
+		  userId: 123,
+		  username: 'testuser',
+		  email: 'test@user.com'
+		}
+	  },
+	  400: {
+		description: 'Bad request - could be invalid data, or that the user/email already exists in the database',
+		type: 'object',
+		properties: {
+		  success: { type: 'boolean' },
+		  message: { type: 'string' }
+		},
+		example: {
+		  success: false,
+		  message: 'Username is already taken'
+		}
+	  },
+	  500: {
+		description: 'Server error',
+		type: 'object',
+		properties: {
+		  success: { type: 'boolean' },
+		  message: { type: 'string' }
+		},
+		example: {
+		  success: false,
+		  message: 'Internal server error'
+		}
+	  }
+	}
+  };
+
 module.exports = {
   signupSchema,
   signinSchema,
-  logoutSchema
+  logoutSchema,
+  googleSchema
 };
